Move getPositionDifference method from PositionUtils to NarrationUtils and rename it
