+++ common/src/main/java/com/github/khanshoaib3/minecraft_access/utils/position/PositionUtils.java
@@ -3,71 +3,16 @@ package com.github.khanshoaib3.minecraft_access.utils.position;
 import com.github.khanshoaib3.minecraft_access.MainClass;
 import com.github.khanshoaib3.minecraft_access.utils.NarrationUtils;
 import com.github.khanshoaib3.minecraft_access.utils.WorldUtils;
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.resource.language.I18n;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Direction;
 import net.minecraft.util.math.Vec3d;
 
 /**
  * Functions about getting blocks' position.
  */
 public class PositionUtils {
-
-    public static String getPositionDifference(BlockPos blockPos) {
-        MinecraftClient minecraftClient = MinecraftClient.getInstance();
-        if (minecraftClient == null) return "up";
-        if (minecraftClient.player == null) return "up";
-
-        Direction dir = minecraftClient.player.getHorizontalFacing();
-
-//        Vec3d diff = minecraftClient.player.getEyePos().subtract(Vec3d.ofCenter(blockPos)); // post 1.18
-        Vec3d diff = new Vec3d(minecraftClient.player.getX(), minecraftClient.player.getEyeY(), minecraftClient.player.getZ()).subtract(Vec3d.ofCenter(blockPos)); // pre 1.18
-        BlockPos diffBlockPos = new BlockPos((int) diff.x, (int) diff.y, (int) diff.z); // post 1.20
-//        BlockPos diffBlockPos = new BlockPos(Math.round(diff.x), Math.round(diff.y), Math.round(diff.z));
-
-        String diffXBlockPos = "";
-        String diffYBlockPos = "";
-        String diffZBlockPos = "";
-
-        if (diffBlockPos.getX() != 0) {
-            if (dir == Direction.NORTH) {
-                diffXBlockPos = getDifferenceString(diffBlockPos.getX(), "right", "left");
-            } else if (dir == Direction.SOUTH) {
-                diffXBlockPos = getDifferenceString(diffBlockPos.getX(), "left", "right");
-            } else if (dir == Direction.EAST) {
-                diffXBlockPos = getDifferenceString(diffBlockPos.getX(), "away", "behind");
-            } else if (dir == Direction.WEST) {
-                diffXBlockPos = getDifferenceString(diffBlockPos.getX(), "behind", "away");
-            }
-        }
-
-        if (diffBlockPos.getY() != 0) {
-            diffYBlockPos = getDifferenceString(diffBlockPos.getY(), "up", "down");
-        }
-
-        if (diffBlockPos.getZ() != 0) {
-            if (dir == Direction.SOUTH) {
-                diffZBlockPos = getDifferenceString(diffBlockPos.getZ(), "away", "behind");
-            } else if (dir == Direction.NORTH) {
-                diffZBlockPos = getDifferenceString(diffBlockPos.getZ(), "behind", "away");
-            } else if (dir == Direction.EAST) {
-                diffZBlockPos = getDifferenceString(diffBlockPos.getZ(), "right", "left");
-            } else if (dir == Direction.WEST) {
-                diffZBlockPos = getDifferenceString(diffBlockPos.getZ(), "left", "right");
-            }
-        }
-
-        String text;
-        if (dir == Direction.NORTH || dir == Direction.SOUTH)
-            text = String.format("%s  %s  %s", diffZBlockPos, diffYBlockPos, diffXBlockPos);
-        else
-            text = String.format("%s  %s  %s", diffXBlockPos, diffYBlockPos, diffZBlockPos);
-        return text;
-    }
 
     public static String getPosition(BlockPos blockPos) {
         try {
@@ -79,10 +24,6 @@ public class PositionUtils {
             MainClass.errorLog("An error occurred when getting block position.", e);
         }
         return "";
-    }
-
-    public static String getDifferenceString(int blocks, String key1, String key2) {
-        return I18n.translate("minecraft_access.util.position_difference_" + (blocks < 0 ? key1 : key2), Math.abs(blocks));
     }
 
     /**
